requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0
click>=8.1.0
pyyaml>=6.0
